Refactor: Improve code modularity and separation of concerns- Moved display_results() from stock_predictors.py to app.py- Moved update_selected_tickers() from stock_predictors.py to app.py- Moved train_model() from stock_predictors.py to model.py- Created new pipeline.py and moved execute_pipeline() to it- Updated imports and removed redundant code- Ensured no changes to existing function logic
